requirements.txt
pymongo
playwright
asyncio
bs4
requests
python-dotenv
google-genai